+++ build.sh
@@ -7,6 +7,9 @@ set -eo pipefail -o xtrace
 # Update this line accordingly if you are not building *from* darwin-x86_64 or linux-x86_64
 export PATH=$PATH:$ANDROID_NDK_HOME/toolchains/llvm/prebuilt/`uname | tr '[:upper:]' '[:lower:]'`-x86_64/bin
 
+# Required for 'ring' dependency to cross-compile to Android platform, must be at least 21
+export CFLAGS="-D__ANDROID_API__=21"
+
 # IMPORTANT: make sure every target is not a substring of a different one. We check for them with grep later on
 BUILD_TARGETS="${BUILD_TARGETS:-aarch64,armv7,x86_64,i686}"