+++ CHANGELOG.md
@@ -29,8 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Return Transaction details data list from list_transactions
 - Add android unit tests
 - Add android tests
-- Update `bdk` to rev `33a5ba6`
+- Update `bdk` to rev `0.4`
 - Add multi-thread coroutine test, fix Lib.call() request string handling
-- Pin `ring` version to 0.16.15
+- Add create and restore extended key APIs
 
 [unreleased]: https://github.com/bitcoindevkit/bdk-jni/compare/d08725cc...HEAD